fix(backup): 4 temuan compiler CI — Float cast, sealed override, smart-cast, import auth
DeltaEncoder.decode: accuracy/speed ke Float eksplisit (ACC/SPEED_SCALE
Double). BackupSource: waypoints jadi abstract val yang dioverride —
lambda sources.map di backupSources tidak bisa resolve properti
per-subclass tanpa itu. Restore passes: getById(row.source*Id!!) —
pass filter tidak membawa smart-cast ke try blok. BackupRepository:
import gotrue.auth untuk currentUserOrNull().

+++ app/src/main/java/com/nyasar/app/backup/DeltaEncoder.kt
@@ -183,8 +183,8 @@ object DeltaEncoder {
                         lat = latCm / COORD_SCALE,
                         lon = lonCm / COORD_SCALE,
                         elevationM = if (eleNull) null else prevEleCm / ELE_SCALE,
-                        speedMps = if (speedNull) null else speedCms / SPEED_SCALE,
-                        accuracyMeters = accDm / ACC_SCALE,
+                        speedMps = if (speedNull) null else (speedCms / SPEED_SCALE).toFloat(),
+                        accuracyMeters = (accDm / ACC_SCALE).toFloat(),
                         timestampMs = prevTime
                     )
                 )
